Improve efficiency of UResourceDataEntry and UResourcBundle structs.
- We can reorder the fields to allow for better packing.

- Additionally, the comment on fNameBuffer is incorrect. There is no
longer any free space due to padding, as the sizeof(ResourceData) is
now 64 bytes, so it is already aligned, and thus no padding is needed.

Using a char array with size 3 actually wastes space.
It wastes 1 byte on 32-bit and 5 bytes on 64-bit.

Instead we can use a char array of size 4, and due to the uint32_t which
occurs next will be aligned on both and doesn't need padding.